requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.11.4
python-dotenv==1.0.1

langchain
langchain-core
langchain-community
langchain-google-genai

faiss-cpu==1.8.0.post1
pypdf==4.3.1
docx2txt==0.8
tiktoken==0.7.0
numpy==1.26.4
